Decouple core API from so-called core module
