+++ api/src/main/java/com/homeaway/streamingplatform/api/exception/StreamDeletionException.java
@@ -13,23 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.homeaway.streamingplatform.exceptions;
+package com.homeaway.streamingplatform.api.exception;
 
-public class SchemaManagerException extends RuntimeException {
+public class StreamDeletionException extends BaseStreamException {
+    private static final long serialVersionUID = -5793888606557229357L;
 
-
-    public SchemaManagerException() {
-    }
-
-    public SchemaManagerException(String message) {
-        super(message);
-    }
-
-    public SchemaManagerException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public SchemaManagerException(Throwable cause) {
-        super(cause);
+    public StreamDeletionException(String streamName) {
+        super(streamName);
     }
 }
