+++ api/src/main/java/com/homeaway/streamingplatform/api/model/StreamClient.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.homeaway.streamingplatform.model;
+package com.homeaway.streamingplatform.api.model;
 
 import java.util.List;
 import java.util.Objects;
