+++ core/pom.xml
@@ -27,6 +27,12 @@
 
         <dependency>
             <groupId>com.homeaway.streamingplatform</groupId>
+            <artifactId>stream-registry-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.homeaway.streamingplatform</groupId>
             <artifactId>infra-provider</artifactId>
             <version>${project.version}</version>
         </dependency>
