+++ core/src/main/java/com/homeaway/streamingplatform/extensions/validation/EmptyStreamValidator.java
@@ -19,8 +19,8 @@ import java.util.Map;
 
 import lombok.extern.slf4j.Slf4j;
 
-import com.homeaway.streamingplatform.exceptions.InvalidStreamException;
-import com.homeaway.streamingplatform.model.Stream;
+import com.homeaway.streamingplatform.api.exception.InvalidStreamException;
+import com.homeaway.streamingplatform.api.model.Stream;
 
 @SuppressWarnings("unused")
 @Slf4j
