+++ pom.xml
@@ -7,6 +7,7 @@
     <name>${project.groupId}:${project.artifactId}</name>
     <modules>
         <module>schema-library-events</module>
+        <module>api</module>
         <module>infra-provider</module>
         <module>infra-provider-kafka</module>
         <module>core</module>
